requirements.txt
psycopg2-binary==2.9.9
requests==2.32.3
pytz==2024.1
python-dotenv==1.0.1
